Remove legacy refraction path
Migrate the remaining transmission scenes to frame-graph scene-texture transmission, remove duplicate Scene 142 assets, and refresh bundle/parity metadata.
