--- lab/babylon-ref-scene142.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8" />
-  <title>Babylon.js Reference — Scene 142: KHR_materials_volume_testing Transmission</title>
-  <style>
-    html, body { margin: 0; padding: 0; width: 100%; height: 100%; overflow: hidden; background: #000; }
-    canvas { width: 100%; height: 100%; display: block; }
-  </style>
-</head>
-<body>
-  <canvas id="renderCanvas"></canvas>
-  <script type="module" src="/src/bjs/scene142.ts"></script>
-</body>
-</html>
