--- lab/bundle-bjs-scene142.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8" />
-  <title>Babylon.js — Scene 142 (Bundle)</title>
-  <style>
-    html, body { margin: 0; padding: 0; width: 100%; height: 100%; overflow: hidden; background: #000; }
-    canvas { width: 100%; height: 100%; display: block; }
-  </style>
-</head>
-<body>
-  <canvas id="renderCanvas"></canvas>
-  <script type="module" src="/bundle/bjs-scene142.js"></script>
-</body>
-</html>
