--- lab/bundle-scene142.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8" />
-  <title>Scene 142 — KHR_materials_volume_testing Transmission (Bundle)</title>
-  <style>
-    body { margin: 0; overflow: hidden; background: #000; }
-    canvas { display: block; width: 100vw; height: 100vh; }
-  </style>
-</head>
-<body>
-  <canvas id="renderCanvas" width="1280" height="720"></canvas>
-  <script src="/loader.js"></script>
-  <script type="module" src="/bundle/scene142.js"></script>
-</body>
-</html>
