+++ lab/index.html
@@ -2900,7 +2900,7 @@
                             <td>KHR_materials_transmission</td>
                             <td>⚡</td>
                             <td>✅</td>
-                            <td>Legacy env-only refraction (Scene 30); frame-graph scene-texture transmission workbench (Scene 142)</td>
+                            <td>Frame-graph scene-texture transmission (Scenes 30/33/112)</td>
                         </tr>
                         <tr>
                             <td>KHR_materials_ior</td>
