--- lab/src/bjs/scene142.ts
@@ -1,50 +1,0 @@
-import { ArcRotateCamera } from "@babylonjs/core/Cameras/arcRotateCamera";
-import { WebGPUEngine } from "@babylonjs/core/Engines/webgpuEngine";
-import "@babylonjs/core/Helpers/sceneHelpers";
-import "@babylonjs/core/Loading/loadingScreen";
-import { SceneLoader } from "@babylonjs/core/Loading/sceneLoader";
-import { Vector3 } from "@babylonjs/core/Maths/math.vector";
-import { Scene } from "@babylonjs/core/scene";
-import "@babylonjs/loaders/glTF";
-
-(async function () {
-    const __initStart = performance.now();
-    const canvas = document.getElementById("renderCanvas") as HTMLCanvasElement;
-    const engine = new WebGPUEngine(canvas, { antialias: true, adaptToDeviceRatio: true });
-    await engine.initAsync();
-
-    const scene = new Scene(engine);
-    (window as any).scene = scene;
-
-    await SceneLoader.AppendAsync("https://assets.babylonjs.com/meshes/", "KHR_materials_volume_testing.glb", scene);
-
-    scene.createDefaultEnvironment({ createGround: false, createSkybox: false });
-
-    // Exact camera values captured from playground #YG3BBF#16 (match lite scene142).
-    const cam = new ArcRotateCamera("camera", Math.PI / 2, Math.PI / 2, 1.1856086997830126, new Vector3(-0.2914360649171073, 0.4, 0.3975263311541397), scene);
-    cam.fov = 0.8;
-    cam.minZ = 0.0697417;
-    cam.maxZ = 6974.17;
-    cam.attachControl(canvas, true);
-    scene.activeCamera = cam;
-
-    const eng = engine as unknown as { _drawCalls?: { fetchNewFrame(): void; current: number } };
-    scene.onBeforeRenderObservable.add(() => {
-        eng._drawCalls?.fetchNewFrame();
-    });
-    scene.onAfterRenderObservable.add(() => {
-        canvas.dataset.drawCalls = String(eng._drawCalls?.current ?? 0);
-    });
-    await scene.whenReadyAsync();
-    engine.runRenderLoop(() => scene.render());
-    window.addEventListener("resize", () => engine.resize());
-
-    await new Promise<void>((resolve) => scene.onAfterRenderObservable.addOnce(() => resolve()));
-    canvas.dataset.camAlpha = String(cam.alpha);
-    canvas.dataset.camBeta = String(cam.beta);
-    canvas.dataset.camRadius = String(cam.radius);
-    canvas.dataset.camTarget = `${cam.target.x},${cam.target.y},${cam.target.z}`;
-    canvas.dataset.camFov = String(cam.fov);
-    canvas.dataset.initMs = String(performance.now() - __initStart);
-    canvas.dataset.ready = "true";
-})().catch(console.error);
