+++ lab/src/lite/scene33.ts
@@ -1,20 +1,9 @@
 // Scene 33 — KHR_lights_punctual glTF test — matches Babylon #YG3BBF#54
 // Loads LightsPunctualLamp.glb (KHR_lights_punctual + KHR_materials_transmission),
-// default environment (IBL only), default camera flipped by +π. This scene stays
-// on the legacy env-only refraction path; Scene 142 is the frame-graph
-// scene-texture transmission workbench.
+// default environment (IBL only), default camera flipped by +π, and frame-graph
+// scene-texture transmission enabled.
 
-import { addToScene, startEngine, createEngine, createSceneContext, createDefaultCamera, loadEnvironment, loadGltf, attachControl, registerScene } from "babylon-lite";
-
-function disableTransmissiveMaterials(entity: unknown): void {
-    const node = entity as { material?: { transmissive?: boolean }; children?: readonly unknown[] };
-    if (node.material) {
-        node.material.transmissive = false;
-    }
-    for (const child of node.children ?? []) {
-        disableTransmissiveMaterials(child);
-    }
-}
+import { addToScene, startEngine, createEngine, createSceneContext, createDefaultCamera, loadEnvironment, loadGltf, attachControl, registerScene, getFrameGraph, type RenderTask } from "babylon-lite";
 
 async function main(): Promise<void> {
     const __initStart = performance.now();
@@ -22,11 +11,9 @@ async function main(): Promise<void> {
 
     const engine = await createEngine(canvas);
     const scene = createSceneContext(engine);
+    (getFrameGraph(scene)._tasks[0] as RenderTask)._config.transmission = { copyCount: 1 };
 
     const asset = await loadGltf(engine, "https://raw.githubusercontent.com/KhronosGroup/glTF-Sample-Assets/main/Models/LightsPunctualLamp/glTF-Binary/LightsPunctualLamp.glb");
-    for (const entity of asset.entities) {
-        disableTransmissiveMaterials(entity);
-    }
     addToScene(scene, asset);
 
     await loadEnvironment(scene, "https://assets.babylonjs.com/environments/environmentSpecular.env", {
