--- tests/parity/scenes/scene142-volume-transmission.spec.ts
@@ -1,35 +1,0 @@
-/**
- * Scene 142 — KHR_materials_volume_testing with scene-texture transmission.
- *
- * Copy of Scene 30 used as the frame-graph transmission workbench.
- */
-import { test, expect } from "@playwright/test";
-import * as path from "path";
-import { attachCompareArtifacts, captureGolden, compareImages, getSceneConfig } from "../compare-utils";
-
-const sceneConfig = getSceneConfig(142);
-const REFERENCE_DIR = path.resolve(__dirname, "../../../reference/scene142-volume-transmission");
-const GOLDEN_REF = path.join(REFERENCE_DIR, "babylon-ref-golden.png");
-
-test.skip(!!sceneConfig.skipParity, "Scene 142 skipped via skipParity in scene-config.json");
-
-test("Scene 142 — KHR_materials_volume_testing transmission matches Babylon.js reference", async ({ page }, testInfo) => {
-    const browser = page.context().browser()!;
-    await captureGolden(browser, { sceneId: 142 });
-
-    await page.goto("/scene142.html");
-    await page.waitForFunction(() => document.querySelector("canvas")?.dataset.ready === "true", { timeout: 30_000 });
-    await page.waitForTimeout(1000);
-
-    const screenshotPath = path.join(REFERENCE_DIR, "test-actual.png");
-    await page.locator("canvas").screenshot({ path: screenshotPath });
-
-    const full = compareImages(screenshotPath, GOLDEN_REF);
-    await attachCompareArtifacts(testInfo, screenshotPath, GOLDEN_REF, REFERENCE_DIR);
-    console.log(`Full image (${full.totalPixels} px):`);
-    console.log(`  MAD: ${full.mad.toFixed(3)}`);
-    console.log(`  Exact: ${((100 * full.exactMatch) / full.totalPixels).toFixed(1)}%`);
-    console.log(`  ≤5: ${((100 * full.within5) / full.totalPixels).toFixed(1)}%`);
-
-    expect(full.mad, `Full image MAD should be ≤ ${sceneConfig.maxMad}`).toBeLessThanOrEqual(sceneConfig.maxMad);
-});
